Backend/requirements.txt
Flask == 3.0.3
requests == 2.31.0
pytesseract == 0.3.10
Pillow == 10.3.0
Werkzeug == 3.0.3
python-docx == 1.1.2
flask_cors == 4.0.1
PyMuPDF == 1.24.2
khmer-nltk == 1.6
pdf2image == 1.17.0
PyPDF2 == 3.0.1
nltk == 3.8.1
pyspellchecker == 0.8.3
docx2pdf == 0.1.8
gunicorn == 23.0.0